pyproject.toml
[project]
name = "ecolab-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.14",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.4",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
]
